Set up test package
