+++ test/mock_comment.py
@@ -1,0 +1,5 @@
+import praw
+
+class Comment(praw.models.Comment):
+    def __init__(self):
+        print(123)
